requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
